Fix AnonymousGlobalsBench
